backend/requirements.txt

uvicorn[standard]>=0.24.0

# AI and MCP SDKs
openai>=1.12.0
mcp>=0.1.0

# Database
sqlmodel~=0.0.16

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# CORS
fastapi[all]>=0.104.1

# Development dependencies
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.2

# PostgreSQL driver
psycopg2-binary>=2.9.9